Make the SOA_SERIAL field a BigInteger

+++ netbox_dns/migrations/0006_zone_soa_serial_auto.py
@@ -20,7 +20,7 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="zone",
             name="soa_serial",
-            field=models.PositiveIntegerField(
+            field=models.BigIntegerField(
                 blank=True,
                 null=True,
                 validators=[
